fix: save return 값을 int 로 casting으로 변경
H2 DB에서 return 값이 Interger입니다.

+++ src/main/java/kr/codesqaud/cafe/post/PostController.java
@@ -36,7 +36,7 @@ public class PostController {
     }
 
     @GetMapping("/posts/{postId}")
-    public String showPostPage(Model model, @PathVariable Long postId) {
+    public String showPostPage(Model model, @PathVariable int postId) {
         Post post = postService.findById(postId);
         model.addAttribute(post);
         return "/post/postDetail";
